+++ src/components/breadcrumb.rs
@@ -48,7 +48,7 @@ impl Component for Breadcrumb {
     fn view(&self) -> Html {
         let mut classes = Classes::from("breadcrumb");
         if let Some(extra) = &self.props.classes {
-            classes = classes.extend(extra);
+            classes.push(extra);
         }
         if let Some(size) = &self.props.size {
             classes.push(&size.to_string());