+++ src/elements/content.rs
@@ -40,7 +40,7 @@ impl Component for Content {
     fn view(&self) -> Html {
         let mut classes = Classes::from("content");
         if let Some(extra) = &self.props.classes {
-            classes = classes.extend(extra);
+            classes.push(extra);
         }
         let tag = self.props.tag.clone();
         html! {