+++ src/elements/image.rs
@@ -39,7 +39,7 @@ impl Component for Image {
     fn view(&self) -> Html {
         let mut classes = Classes::from("image");
         if let Some(extra) = &self.props.classes {
-            classes = classes.extend(extra);
+            classes.push(extra);
         }
         if let Some(size) = &self.props.size {
             classes.push(&size.to_string());